+++ main/java/fr/efrei/repository/IGymAdminRepository.java
